+++ js/data.js
@@ -44,4 +44,4 @@ const LNG_MAX = 139.8;
 
 const SIMILAR_COUNT = 4;
 
-export {TYPES,TITLES,DESCRIPTION,FEATURES,PHOTOS,TIME_CHECKS,LAT_MIN,LAT_MAX,LNG_MIN,LNG_MAX,SIMILAR_COUNT};
+export {TYPES,TITLES, DESCRIPTION, FEATURES, PHOTOS, TIME_CHECKS, LAT_MIN, LAT_MAX, LNG_MIN, LNG_MAX, SIMILAR_COUNT};
